fix: resolve GCC 15 compatibility issues across all platforms
- BlockMemoryPool::AlignBlockSize: avoid Max<T> template entirely to prevent
  type deduction failures (uint32_t vs size_t vs unsigned long) on 32-bit ARM,
  RISC-V, and 64-bit Windows/Linux
- test/stktest.cpp: wrap GetTls/SetTls in namespace block (GCC 15 rejects
  qualified-name definitions outside their namespace)
- test/generic/stktest_kernelservice.cpp: cast Ticks to int32_t in CHECK_EQUAL
  to avoid CppUTest StringFrom ambiguity with int64_t on ARM GCC 15

+++ test/generic/stktest_kernelservice.cpp
@@ -131,8 +131,8 @@ TEST(KernelService, GetTickResolution)
     kernel.AddTask(&task);
     kernel.Start();
 
-    CHECK_EQUAL(periodicity, g_KernelService->GetTickResolution());
-    CHECK_EQUAL(periodicity, stk::GetTickResolution());
+    CHECK_EQUAL((int32_t)periodicity, (int32_t)g_KernelService->GetTickResolution());
+    CHECK_EQUAL((int32_t)periodicity, (int32_t)stk::GetTickResolution());
 }
 
 TEST(KernelService, GetTicks)
